sandbox_externalkey.go: split for cross compilation
runc/libcontainer split the `State` struct into platform specific structs
in
https://github.com/opencontainers/runc/commit/fe1cce69b31faee84a17abfabbac079adea53f06.
As a result, `NamespacePaths` isn't anymore in a global struct and
libnetwork is not cross-compiling in Docker (specifically on Windows) because
`sandbox_externalkey.go` is using `NamespacePaths`.
This patch splits `sandbox_externalkey.go` into platform specific
files and moves common things to a generic `sandbox_externalkey.go`.
